refactor: Break BlockDefinitionCompiler into several files.
